refactor copyCourse.js
