+++ step_definitions/homework-steps.js
@@ -3,7 +3,6 @@ const path = require('path');
 
 const loginPage = require('../page-objects/pages/generalPagesBeforeLogin/LoginPage.js');
 const startPage = require('../page-objects/pages/generalPagesBeforeLogin/StartPageBeforeLogin.js');
-const copyCourse = require('../page-objects/copyCourse');
 const homework = require('../page-objects/homework');
 const hMWRKAddEditHomeworkPage = require('../page-objects/pages/HMWRKAddEditHomeworkPage.js');
 const hMWRKHomeworkListPage = require('../page-objects/pages/HMWRKHomeworkListPage');
@@ -98,7 +97,7 @@ Then(
 Given(/^the Teacher creates one course with (.*) and pupil with:$/, function (
 	coursename
 ) {
-	return copyCourse.create(coursename);
+	//return copyCourse.create(coursename);
 });
 When(/^Teacher creates a homework for the course (.*)$/, function (coursename) {
 	return homework.clickCreateNewTaskInTheCourse(coursename);
